Refactor clean_yum_metadata to be separate from pkghandler.py
