+++ convert2rhel/pkghandler.py
@@ -1079,28 +1079,6 @@ def _get_packages_to_update_dnf(reposdir):
     return packages
 
 
-def clean_yum_metadata():
-    """Remove cached metadata from yum.
-
-    This is to make sure that Convert2RHEL works with up-to-date data from repositories before, for instance, querying
-    whether the system has the latest package versions installed, or before checking whether enabled repositories have
-    accessible URLs.
-    """
-    # We are using run_subprocess here as an alternative to call_yum_cmd
-    # which doesn't apply the correct --enablerepos option because at this point
-    # the tool didn't initialize the necessary functions in SystemInfo.
-    # The viable solution was calling the yum command as a subprocess manually
-    # instead of using that function wrapper.
-    output, ret_code = utils.run_subprocess(("yum", "clean", "metadata", "--quiet"), print_output=False)
-    loggerinst.debug("Output of yum clean metadata:\n%s" % output)
-
-    if ret_code != 0:
-        loggerinst.warning("Failed to clean yum metadata:\n%s" % output)
-        return
-
-    loggerinst.info("Cached yum metadata cleaned successfully.")
-
-
 def _package_version_cmp(pkg_1, pkg_2):
     """Compare the version key in a given package name.
