+++ convert2rhel/pkgmanager/__init__.py
@@ -15,6 +15,12 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
+import logging
+
+from convert2rhel import utils
+
+
+loggerinst = logging.getLogger(__name__)
 
 try:
     from yum import *
@@ -63,3 +69,27 @@ def create_transaction_handler():
         from convert2rhel.pkgmanager.handlers.dnf import DnfTransactionHandler
 
         return DnfTransactionHandler()
+
+
+def clean_yum_metadata():
+    """Remove cached metadata from yum.
+
+    This is to make sure that Convert2RHEL works with up-to-date data from repositories before, for instance, querying
+    whether the system has the latest package versions installed, or before checking whether enabled repositories have
+    accessible URLs.
+    """
+    # We are using run_subprocess here as an alternative to call_yum_cmd
+    # which doesn't apply the correct --enablerepos option because if we call this
+    # earlier in the conversion the tool doesn't initialize the necessary functions in SystemInfo.
+    # The viable solution was calling the yum command as a subprocess manually
+    # instead of using that function wrapper.
+    output, ret_code = utils.run_subprocess(
+        ("yum", "clean", "metadata", "--enablerepo=*", "--quiet"), print_output=False
+    )
+    loggerinst.debug("Output of yum clean metadata:\n%s" % output)
+
+    if ret_code != 0:
+        loggerinst.warning("Failed to clean yum metadata:\n%s" % output)
+        return
+
+    loggerinst.info("Cached repositories metadata cleaned successfully.")
