+++ convert2rhel/unit_tests/main_test.py
@@ -404,7 +404,7 @@ def test_main(monkeypatch):
     monkeypatch.setattr(system_info, "resolve_system_info", resolve_system_info_mock)
     monkeypatch.setattr(breadcrumbs, "collect_early_data", collect_early_data_mock)
     monkeypatch.setattr(pkghandler, "clear_versionlock", clear_versionlock_mock)
-    monkeypatch.setattr(pkghandler, "clean_yum_metadata", clean_yum_metadata_mock)
+    monkeypatch.setattr(pkgmanager, "clean_yum_metadata", clean_yum_metadata_mock)
     monkeypatch.setattr(redhatrelease.system_release_file, "backup", system_release_file_mock)
     monkeypatch.setattr(redhatrelease.os_release_file, "backup", os_release_file_mock)
     monkeypatch.setattr(repo, "backup_yum_repos", backup_yum_repos_mock)
@@ -494,7 +494,7 @@ def test_main_rollback_pre_ponr_changes_phase(monkeypatch):
     monkeypatch.setattr(system_info, "resolve_system_info", resolve_system_info_mock)
     monkeypatch.setattr(breadcrumbs, "collect_early_data", collect_early_data_mock)
     monkeypatch.setattr(pkghandler, "clear_versionlock", clear_versionlock_mock)
-    monkeypatch.setattr(pkghandler, "clean_yum_metadata", clean_yum_metadata_mock)
+    monkeypatch.setattr(pkgmanager, "clean_yum_metadata", clean_yum_metadata_mock)
     monkeypatch.setattr(checks, "perform_system_checks", perform_system_checks_mock)
     monkeypatch.setattr(redhatrelease.system_release_file, "backup", system_release_file_mock)
     monkeypatch.setattr(redhatrelease.os_release_file, "backup", os_release_file_mock)
@@ -552,7 +552,7 @@ def test_main_rollback_post_ponr_changes_phase(monkeypatch, caplog):
     monkeypatch.setattr(system_info, "resolve_system_info", resolve_system_info_mock)
     monkeypatch.setattr(breadcrumbs, "collect_early_data", collect_early_data_mock)
     monkeypatch.setattr(pkghandler, "clear_versionlock", clear_versionlock_mock)
-    monkeypatch.setattr(pkghandler, "clean_yum_metadata", clean_yum_metadata_mock)
+    monkeypatch.setattr(pkgmanager, "clean_yum_metadata", clean_yum_metadata_mock)
     monkeypatch.setattr(checks, "perform_system_checks", perform_system_checks_mock)
     monkeypatch.setattr(redhatrelease.system_release_file, "backup", system_release_file_mock)
     monkeypatch.setattr(redhatrelease.os_release_file, "backup", os_release_file_mock)
