+++ convert2rhel/unit_tests/repo_test.py
@@ -18,10 +18,14 @@
 import os
 
 import pytest
+import six
 
 from convert2rhel import repo
-from convert2rhel.systeminfo import system_info
 from convert2rhel.unit_tests.conftest import all_systems, centos8
+
+
+six.add_move(six.MovedModule("mock", "mock", "unittest.mock"))
+from six.moves import mock
 
 
 @pytest.mark.parametrize(
